requirements.txt
<<<<<<< HEAD
alembic==1.11.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
SQLAlchemy==2.0.15
typing_extensions==4.6.3
Werkzeug==2.3.4
=======
alembic==1.11.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
SQLAlchemy==2.0.15
typing_extensions==4.6.3
Werkzeug==2.3.4
>>>>>>> 230b274459de85cb813a0696f2c26ab38768a098
